+++ javascript/node/test/element_finding_test.js
@@ -17,9 +17,13 @@
 
 require('./lib/_bootstrap')(module);
 
-var assert = require('assert'),
+var fail = require('assert').fail,
     By = require('selenium-webdriver').By,
-    error = require('selenium-webdriver').error;
+    error = require('selenium-webdriver').error,
+    assertThat = require('selenium-webdriver/testing/asserts').assertThat,
+    equals = require('selenium-webdriver/testing/asserts').equals,
+    greaterThan = require('selenium-webdriver/testing/asserts').greaterThan,
+    startsWith = require('selenium-webdriver/testing/asserts').startsWith;
 
 var test = require('./lib/testbase'),
     Browser = test.Browser,
@@ -54,8 +58,8 @@ test.suite(function(env) {
       test.it('should fail if ID not present on page', function() {
         driver.get(Pages.formPage);
         driver.findElement(By.id('nonExistantButton')).
-            then(assert.fail, function(e) {
-              assert.equal(error.ErrorCode.NO_SUCH_ELEMENT, e.code);
+            then(fail, function(e) {
+              assertThat(e.code, equals(error.ErrorCode.NO_SUCH_ELEMENT));
             });
       });
 
@@ -65,7 +69,7 @@ test.suite(function(env) {
           function() {
             driver.get(Pages.nestedPage);
             driver.findElements(By.id('2')).then(function(elements) {
-              assert.equal(elements.length, 8);
+              assertThat(elements.length, equals(8));
             });
           });
     });
@@ -86,21 +90,17 @@ test.suite(function(env) {
 
       test.it('should work when link text contains equals sign', function() {
         driver.get(Pages.xhtmlTestPage);
-        driver.findElement(By.linkText('Link=equalssign')).
-            getAttribute('id').
-            then(function(id) {
-              assert.equal('linkWithEqualsSign', id);
-            });
+        var id = driver.findElement(By.linkText('Link=equalssign')).
+            getAttribute('id');
+        assertThat(id, equals('linkWithEqualsSign'));
       });
 
       test.it('matches by partial text when containing equals sign',
         function() {
           driver.get(Pages.xhtmlTestPage);
-          driver.findElement(By.partialLinkText('Link=')).
-              getAttribute('id').
-              then(function(id) {
-                assert.equal('linkWithEqualsSign', id);
-              });
+          var id = driver.findElement(By.partialLinkText('Link=')).
+              getAttribute('id');
+          assertThat(id, equals('linkWithEqualsSign'));
         });
 
       test.it('works when searching for multiple and text contains =',
@@ -108,11 +108,11 @@ test.suite(function(env) {
             driver.get(Pages.xhtmlTestPage);
             driver.findElements(By.linkText('Link=equalssign')).
                 then(function(elements) {
-                  assert.equal(elements.length, 1);
+                  assertThat(elements.length, equals(1));
                   return elements[0].getAttribute('id');
                 }).
                 then(function(id) {
-                  assert.equal('linkWithEqualsSign', id);
+                  assertThat(id, equals('linkWithEqualsSign'));
                 });
           });
 
@@ -122,20 +122,20 @@ test.suite(function(env) {
             driver.get(Pages.xhtmlTestPage);
             driver.findElements(By.partialLinkText('Link=')).
                 then(function(elements) {
-                  assert.equal(elements.length, 1);
+                  assertThat(elements.length, equals(1));
                   return elements[0].getAttribute('id');
                 }).
                 then(function(id) {
-                  assert.equal('linkWithEqualsSign', id);
+                  assertThat(id, equals('linkWithEqualsSign'));
                 });
-          });
+      });
 
       test.it('should be able to find multiple exact matches',
           function() {
             driver.get(Pages.xhtmlTestPage);
             driver.findElements(By.linkText('click me')).
                 then(function(elements) {
-                  assert.equal(elements.length, 2);
+                  assertThat(elements.length, equals(2));
                 });
           });
 
@@ -144,33 +144,31 @@ test.suite(function(env) {
             driver.get(Pages.xhtmlTestPage);
             driver.findElements(By.partialLinkText('ick me')).
                 then(function(elements) {
-                  assert.equal(elements.length, 2);
+                  assertThat(elements.length, equals(2));
                 });
           });
 
       test.ignore(browsers(Browser.OPERA)).
       it('works on XHTML pages', function() {
         driver.get(test.whereIs('actualXhtmlPage.xhtml'));
-        driver.findElement(By.linkText('Foo')).getText().then(function(text) {
-          assert.equal(text, 'Foo');
-        });
+
+        var el = driver.findElement(By.linkText('Foo'));
+        assertThat(el.getText(), equals('Foo'));
       });
     });
 
     describe('By.name()', function() {
       test.it('should work', function() {
         driver.get(Pages.formPage);
-        driver.findElement(By.name('checky')).
-            getAttribute('value').
-            then(function(value) {
-              assert.equal('furrfu', value);
-            });
+
+        var el = driver.findElement(By.name('checky'));
+        assertThat(el.getAttribute('value'), equals('furrfu'));
       });
 
       test.it('should find multiple elements with same name', function() {
         driver.get(Pages.nestedPage);
         driver.findElements(By.name('checky')).then(function(elements) {
-          assert.ok(elements.length > 1);
+          assertThat(elements.length, greaterThan(1));
         });
       });
 
@@ -192,70 +190,58 @@ test.suite(function(env) {
     describe('By.className()', function() {
       test.it('should work', function() {
         driver.get(Pages.xhtmlTestPage);
-        driver.findElement(By.className('extraDiv')).
-            getText().
-            then(function(text) {
-              var expected = 'Another div starts here.';
-              var actual = text.substring(0, expected.length);
-              assert.equal(actual, expected);
-            });
+
+        var el = driver.findElement(By.className('extraDiv'));
+        assertThat(el.getText(), startsWith('Another div starts here.'));
       });
 
       test.it('should work when name is first name among many', function() {
         driver.get(Pages.xhtmlTestPage);
-        driver.findElement(By.className('nameA')).
-            getText().
-            then(function(text) {
-              assert.equal(text, 'An H2 title');
-            });
+
+        var el = driver.findElement(By.className('nameA'));
+        assertThat(el.getText(), equals('An H2 title'));
       });
 
       test.it('should work when name is last name among many', function() {
         driver.get(Pages.xhtmlTestPage);
-        driver.findElement(By.className('nameC')).
-            getText().
-            then(function(text) {
-              assert.equal(text, 'An H2 title');
-            });
+
+        var el = driver.findElement(By.className('nameC'));
+        assertThat(el.getText(), equals('An H2 title'));
       });
 
       test.it('should work when name is middle of many', function() {
         driver.get(Pages.xhtmlTestPage);
-        driver.findElement(By.className('nameBnoise')).
-            getText().
-            then(function(text) {
-              assert.equal(text, 'An H2 title');
-            });
+
+        var el = driver.findElement(By.className('nameBnoise'));
+        assertThat(el.getText(), equals('An H2 title'));
       });
 
       test.it('should work when name surrounded by whitespace', function() {
         driver.get(Pages.xhtmlTestPage);
-        driver.findElement(By.className('spaceAround')).
-            getText().
-            then(function(text) {
-              assert.equal(text, 'Spaced out');
-            });
+
+        var el = driver.findElement(By.className('spaceAround'));
+        assertThat(el.getText(), equals('Spaced out'));
       });
 
       test.it('should fail if queried name only partially matches', function() {
         driver.get(Pages.xhtmlTestPage);
         driver.findElement(By.className('nameB')).
-            then(assert.fail, function(e) {
-              assert.equal(error.ErrorCode.NO_SUCH_ELEMENT, e.code);
+            then(fail, function(e) {
+              assertThat(e.code, equals(error.ErrorCode.NO_SUCH_ELEMENT));
             });
       });
 
       test.it('should be able to find multiple matches', function() {
         driver.get(Pages.xhtmlTestPage);
         driver.findElements(By.className('nameC')).then(function(elements) {
-          assert.ok(elements.length > 1);
+          assertThat(elements.length, greaterThan(1));
         });
       });
 
       test.it('does not permit compound class names', function() {
         driver.get(Pages.xhtmlTestPage);
-        driver.findElement(By.className('a b')).then(assert.fail, pass);
-        driver.findElements(By.className('a b')).then(assert.fail, pass);
+        driver.findElement(By.className('a b')).then(fail, pass);
+        driver.findElements(By.className('a b')).then(fail, pass);
         function pass() {}
       });
     });
@@ -264,7 +250,7 @@ test.suite(function(env) {
       test.it('should work with multiple matches', function() {
         driver.get(Pages.xhtmlTestPage);
         driver.findElements(By.xpath('//div')).then(function(elements) {
-          assert.ok(elements.length > 1, elements.length);
+          assertThat(elements.length, greaterThan(1));
         });
       });
 
@@ -278,16 +264,15 @@ test.suite(function(env) {
     describe('By.tagName()', function() {
       test.it('works', function() {
         driver.get(Pages.formPage);
-        driver.findElement(By.tagName('input')).getTagName().
-            then(function(name) {
-              assert.equal(name, 'input');
-            });
+
+        var el = driver.findElement(By.tagName('input'));
+        assertThat(el.getTagName(), equals('input'));
       });
 
       test.it('can find multiple elements', function() {
         driver.get(Pages.formPage);
         driver.findElements(By.tagName('input')).then(function(elements) {
-          assert.ok(elements.length > 1);
+          assertThat(elements.length, greaterThan(1));
         });
       });
     });
@@ -302,26 +287,31 @@ test.suite(function(env) {
       test.it('can find multiple elements', function() {
         driver.get(Pages.xhtmlTestPage);
         driver.findElements(By.css('p')).then(function(elements) {
-          assert.ok(elements.length > 1);
+          assertThat(elements.length, greaterThan(1));
         });
         // Pass if no error.
       });
 
-      test.it('should be able to find element by compound CSS selector',
-        function() {
-          driver.get(Pages.xhtmlTestPage);
-          assertAttributeIs(
-              driver.findElement(By.css('div.extraDiv, div.content')),
-              'class', 'content');
-        });
+      test.it(
+          'should find first matching element when searching by ' +
+              'compound CSS selector',
+          function() {
+            driver.get(Pages.xhtmlTestPage);
+            var el = driver.findElement(By.css('div.extraDiv, div.content'));
+            assertThat(el.getAttribute('class'), equals('content'));
+          });
 
       test.it('should be able to find multiple elements by compound selector',
           function() {
             driver.get(Pages.xhtmlTestPage);
             driver.findElements(By.css('div.extraDiv, div.content')).
                 then(function(elements) {
-                  assertAttributeIs(elements[0], 'class', 'content');
-                  assertAttributeIs(elements[1], 'class', 'extraDiv');
+                  assertClassIs(elements[0], 'content');
+                  assertClassIs(elements[1], 'extraDiv');
+
+                  function assertClassIs(el, expected) {
+                    assertThat(el.getAttribute('class'), equals(expected));
+                  }
                 });
           });
 
@@ -330,9 +320,9 @@ test.suite(function(env) {
       it('should be able to find element by boolean attribute', function() {
         driver.get(test.whereIs(
             'locators_tests/boolean_attribute_selected.html'));
-        assertAttributeIs(
-            driver.findElement(By.css('option[selected="selected"]')),
-            'value', 'two');
+
+        var el = driver.findElement(By.css('option[selected="selected"]'));
+        assertThat(el.getAttribute('value'), equals('two'));
       });
 
       test.it(
@@ -341,8 +331,9 @@ test.suite(function(env) {
           function() {
             driver.get(test.whereIs(
                 'locators_tests/boolean_attribute_selected.html'));
-            assertAttributeIs(
-                driver.findElement(By.css('option[selected]')), 'value', 'two');
+
+            var el = driver.findElement(By.css('option[selected]'));
+            assertThat(el.getAttribute('value'), equals('two'));
           });
 
       test.it(
@@ -351,15 +342,10 @@ test.suite(function(env) {
           function() {
             driver.get(test.whereIs(
                 'locators_tests/boolean_attribute_selected_html4.html'));
-            assertAttributeIs(driver.findElement(By.css('option[selected]')),
-                'value', 'two');
-          });
 
-      function assertAttributeIs(element, attribute, expected) {
-        element.getAttribute(attribute).then(function(attribute) {
-          assert.equal(attribute, expected);
-        });
-      }
+            var el = driver.findElement(By.css('option[selected]'));
+            assertThat(el.getAttribute('value'), equals('two'));
+          });
     });
   });
 });